Moved: Task tests to new namespace.
